min, max, abs, to_baseN, unfinished is_prime

+++ BigInt_core.cpp
@@ -18,7 +18,7 @@ string BigInt::getBin()
 {
 	if (length < 19) {
         if(sign)
-		    return decimalToBinary(-small);
+		    return decimalToBinary(-(int)small);
         else
             return decimalToBinary(small);
 	}
